test: Setup puppeteer-jest
End to end testing is setup using puppeteer-jest. I've modified the config files I've used inside the config folder. However you will notice a setup.js file inside the root directory. This is required at the moment in order to get puppeteer-jest to work as I've had to modify some globals before the test environment is run. Puppeteer-jest is not configured for jest-configuration per the puppeteer docs. So I've ignored the e2e folder. In order for this configuration to work, headless chrome must be turned ON. I havent ran this in our CI/CD pipeline. I've seen some docs on it, and I'll look more into it.